Backend/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# Backend/db/database.py: 1,2,3
# Backend/models/models.py: 1,2
SQLAlchemy == 1.4.37

# Backend/dependencies.py: 1
# Backend/main.py: 2,3
# Backend/routers/menus.py: 1
# Backend/routers/reservations.py: 1
fastapi == 0.78.0

# Backend/env/lib/python3.10/site-packages/mysql/connector/__init__.py: 34
# Backend/env/lib/python3.10/site-packages/mysql/connector/connection_cext.py: 49,55
# Backend/env/lib/python3.10/site-packages/mysql/connector/cursor_cext.py: 36
# Backend/env/lib/python3.10/site-packages/mysql/connector/django/base.py: 63
# Backend/env/lib/python3.10/site-packages/mysql/connector/django/operations.py: 34
# Backend/env/lib/python3.10/site-packages/mysqlx/protobuf/__init__.py: 72
# Backend/env/lib64/python3.10/site-packages/mysql/connector/__init__.py: 34
# Backend/env/lib64/python3.10/site-packages/mysql/connector/connection_cext.py: 49,55
# Backend/env/lib64/python3.10/site-packages/mysql/connector/cursor_cext.py: 36
# Backend/env/lib64/python3.10/site-packages/mysql/connector/django/base.py: 63
# Backend/env/lib64/python3.10/site-packages/mysql/connector/django/operations.py: 34
# Backend/env/lib64/python3.10/site-packages/mysqlx/protobuf/__init__.py: 72
mysql_connector_python == 8.0.29

# Backend/schemas/schemas.py: 1
pydantic == 1.9.1

# Backend/env/lib/python3.10/site-packages/pyrsistent/_pvector.py: 697
# Backend/env/lib64/python3.10/site-packages/pyrsistent/_pvector.py: 697
pyrsistent == 0.18.1

# Backend/db/database.py: 4
python_dotenv == 0.20.0

# Backend/env/lib/python3.10/site-packages/pkg_resources/tests/data/my-test-package-source/setup.py: 1
# Backend/env/lib64/python3.10/site-packages/pkg_resources/tests/data/my-test-package-source/setup.py: 1
setuptools == 58.1.0

# Backend/main.py: 1
uvicorn == 0.17.6
